fixed all errors
Continue fixing

+++ src/core/stats/rank/Administrator.php
@@ -35,7 +35,7 @@ class Administrator extends Rank {
         return self::STAFF;
     }
 
-    public function getChatTime() : int {
+    public function getChatTime() : float {
         return 0;
     }
 }
